refactor: remove the 'connecting' event
This event was added in https://github.com/socketio/socket.io-client/commit/41956806a7d4da382c043d15efb2abe7c21a24d9

But it does not convey the information that the Socket is actually
sending a CONNECT packet to the server. It should maybe be moved to the
Socket#onopen() method, but let's remove it for now as it is not
documented anywhere.